+++ app/plugins/TagManager/TagManager.php
@@ -539,6 +539,35 @@ class TagManager extends \Piwik\Plugin
         $result[] = 'TagManager_InstallCode';
         $result[] = 'TagManager_InstallCodePublishEnvironmentNote';
         $result[] = 'TagManager_GettingStartedNotice';
+        $result[] = 'TagManager_InvalidDebugUrlError';
+        $result[] = 'TagManager_TagDescriptionHelp';
+        $result[] = 'TagManager_TriggerDescriptionHelp';
+        $result[] = 'TagManager_VariableDescriptionHelp';
+        $result[] = 'TagManager_InstallCodeDataLayerNote';
+        $result[] = 'TagManager_TagsNameDescription';
+        $result[] = 'TagManager_TagsDescriptionDescription';
+        $result[] = 'TagManager_TagsTypeDescription';
+        $result[] = 'TagManager_TagsTriggersDescription';
+        $result[] = 'TagManager_TagsLastUpdatedDescription';
+        $result[] = 'TagManager_TagsActionDescription';
+        $result[] = 'TagManager_TriggersNameDescription';
+        $result[] = 'TagManager_TriggersDescriptionDescription';
+        $result[] = 'TagManager_TriggersTypeDescription';
+        $result[] = 'TagManager_TriggersFilterDescription';
+        $result[] = 'TagManager_TriggersLastUpdatedDescription';
+        $result[] = 'TagManager_TriggersActionDescription';
+        $result[] = 'TagManager_VariablesNameDescription';
+        $result[] = 'TagManager_VariablesDescriptionDescription';
+        $result[] = 'TagManager_VariablesTypeDescription';
+        $result[] = 'TagManager_VariablesLookupTableDescription';
+        $result[] = 'TagManager_VariablesLastUpdatedDescription';
+        $result[] = 'TagManager_VariablesActionDescription';
+        $result[] = 'TagManager_VersionsRevisionDescription';
+        $result[] = 'TagManager_VersionsNameDescription';
+        $result[] = 'TagManager_VersionsDescriptionDescription';
+        $result[] = 'TagManager_VersionsEnvironmentsDescription';
+        $result[] = 'TagManager_VersionsCreatedDescription';
+        $result[] = 'TagManager_VersionsActionDescription';
     }
 
     public function getStylesheetFiles(&$stylesheets)
@@ -550,6 +579,7 @@ class TagManager extends \Piwik\Plugin
         $stylesheets[] = "plugins/TagManager/vue/src/VariableSelectType/VariableSelectType.less";
         $stylesheets[] = "plugins/TagManager/vue/src/Field/FieldVariableTemplate.less";
         $stylesheets[] = "plugins/TagManager/vue/src/ContainerSelector/ContainerSelector.less";
+        $stylesheets[] = "plugins/TagManager/vue/src/ContainerDashboard/ContainerDashboard.less";
         $stylesheets[] = "plugins/TagManager/vue/src/Version/VersionEdit.less";
     }